Fix top add button
