+++ src/app/plugins/kubernetes/components/logging/LoggingIndexPage.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Tabs from 'core/components/tabs/Tabs'
+import Tab from 'core/components/tabs/Tab'
+import LoggingListPage from './LoggingListPage'
+
+const LoggingIndexPage = () => (
+  <Tabs>
+    <Tab value="logging" label="Logging"><LoggingListPage /></Tab>
+  </Tabs>
+)
+
+export default LoggingIndexPage
